chore: move infer_schema and infer_data_type into vortex-dtype
I want to use these in PyVortex without depending on vortex-datafusion.
