+++ vortex-array/src/arrow/mod.rs
@@ -1,5 +1,7 @@
 use vortex_error::VortexResult;
 
+pub use crate::arrow::dtype::{infer_data_type, infer_schema};
+
 mod array;
 mod dtype;
 mod recordbatch;
